separate file drink and food from main
